+++ CHANGES.md
@@ -8,3 +8,4 @@ Changes in 1.1
 - Fix: change school while viewing message in Messages.php
 - Update screenshots
 - Add GetRecipientsHeader(), display current tab in bold in Write.fnc.php
+- Display current tab in bold in Messages.php